requirements.txt
opencv-python>=4.5.0
pyserial>=3.5
openpyxl>=3.0.0
